--- token-lending/program/src/helper.rs
@@ -1,206 +1,0 @@
-use crate::{error::LendingError, state::Reserve};
-use solana_program::{
-    account_info::AccountInfo,
-    clock::Slot,
-    entrypoint::ProgramResult,
-    msg,
-    program::{invoke, invoke_signed},
-    program_error::ProgramError,
-    program_pack::{IsInitialized, Pack},
-    pubkey::Pubkey,
-    sysvar::rent::Rent,
-};
-
-pub fn assert_rent_exempt(rent: &Rent, account_info: &AccountInfo) -> ProgramResult {
-    if !rent.is_exempt(account_info.lamports(), account_info.data_len()) {
-        msg!(&rent.minimum_balance(account_info.data_len()).to_string());
-        Err(LendingError::NotRentExempt.into())
-    } else {
-        Ok(())
-    }
-}
-
-pub fn assert_last_update_slot(reserve: &Reserve, slot: Slot) -> ProgramResult {
-    if !reserve.last_update_slot == slot {
-        Err(LendingError::ReserveStale.into())
-    } else {
-        Ok(())
-    }
-}
-
-pub fn assert_uninitialized<T: Pack + IsInitialized>(
-    account_info: &AccountInfo,
-) -> Result<T, ProgramError> {
-    let account: T = T::unpack_unchecked(&account_info.data.borrow())?;
-    if account.is_initialized() {
-        Err(LendingError::AlreadyInitialized.into())
-    } else {
-        Ok(account)
-    }
-}
-
-/// Unpacks a spl_token `Mint`.
-pub fn unpack_mint(data: &[u8]) -> Result<spl_token::state::Mint, LendingError> {
-    spl_token::state::Mint::unpack(data).map_err(|_| LendingError::InvalidTokenMint)
-}
-
-/// Issue a spl_token `InitializeMint` instruction.
-#[inline(always)]
-pub fn spl_token_init_mint(params: TokenInitializeMintParams<'_, '_>) -> ProgramResult {
-    let TokenInitializeMintParams {
-        mint,
-        rent,
-        authority,
-        token_program,
-        decimals,
-    } = params;
-    let ix = spl_token::instruction::initialize_mint(
-        token_program.key,
-        mint.key,
-        authority,
-        None,
-        decimals,
-    )?;
-    let result = invoke(&ix, &[mint, rent, token_program]);
-    result.map_err(|_| LendingError::TokenInitializeMintFailed.into())
-}
-
-/// Issue a spl_token `InitializeAccount` instruction.
-#[inline(always)]
-pub fn spl_token_init_account(params: TokenInitializeAccountParams<'_>) -> ProgramResult {
-    let TokenInitializeAccountParams {
-        account,
-        mint,
-        owner,
-        rent,
-        token_program,
-    } = params;
-    let ix = spl_token::instruction::initialize_account(
-        token_program.key,
-        account.key,
-        mint.key,
-        owner.key,
-    )?;
-    let result = invoke(&ix, &[account, mint, owner, rent, token_program]);
-    result.map_err(|_| LendingError::TokenInitializeAccountFailed.into())
-}
-
-/// Issue a spl_token `Transfer` instruction.
-#[inline(always)]
-pub fn spl_token_transfer(params: TokenTransferParams<'_, '_>) -> ProgramResult {
-    let TokenTransferParams {
-        source,
-        destination,
-        authority,
-        token_program,
-        amount,
-        authority_signer_seeds,
-    } = params;
-    let result = invoke_signed(
-        &spl_token::instruction::transfer(
-            token_program.key,
-            source.key,
-            destination.key,
-            authority.key,
-            &[],
-            amount,
-        )?,
-        &[source, destination, authority, token_program],
-        &[authority_signer_seeds],
-    );
-    result.map_err(|_| LendingError::TokenTransferFailed.into())
-}
-
-/// Issue a spl_token `MintTo` instruction.
-pub fn spl_token_mint_to(params: TokenMintToParams<'_, '_>) -> ProgramResult {
-    let TokenMintToParams {
-        mint,
-        destination,
-        authority,
-        token_program,
-        amount,
-        authority_signer_seeds,
-    } = params;
-    let result = invoke_signed(
-        &spl_token::instruction::mint_to(
-            token_program.key,
-            mint.key,
-            destination.key,
-            authority.key,
-            &[],
-            amount,
-        )?,
-        &[mint, destination, authority, token_program],
-        &[authority_signer_seeds],
-    );
-    result.map_err(|_| LendingError::TokenMintToFailed.into())
-}
-
-/// Issue a spl_token `Burn` instruction.
-#[inline(always)]
-pub fn spl_token_burn(params: TokenBurnParams<'_, '_>) -> ProgramResult {
-    let TokenBurnParams {
-        mint,
-        source,
-        authority,
-        token_program,
-        amount,
-        authority_signer_seeds,
-    } = params;
-    let result = invoke_signed(
-        &spl_token::instruction::burn(
-            token_program.key,
-            source.key,
-            mint.key,
-            authority.key,
-            &[],
-            amount,
-        )?,
-        &[source, mint, authority, token_program],
-        &[authority_signer_seeds],
-    );
-    result.map_err(|_| LendingError::TokenBurnFailed.into())
-}
-
-pub struct TokenInitializeMintParams<'a: 'b, 'b> {
-    pub mint: AccountInfo<'a>,
-    pub rent: AccountInfo<'a>,
-    pub authority: &'b Pubkey,
-    pub decimals: u8,
-    pub token_program: AccountInfo<'a>,
-}
-
-pub struct TokenInitializeAccountParams<'a> {
-    pub account: AccountInfo<'a>,
-    pub mint: AccountInfo<'a>,
-    pub owner: AccountInfo<'a>,
-    pub rent: AccountInfo<'a>,
-    pub token_program: AccountInfo<'a>,
-}
-
-pub struct TokenTransferParams<'a: 'b, 'b> {
-    pub source: AccountInfo<'a>,
-    pub destination: AccountInfo<'a>,
-    pub amount: u64,
-    pub authority: AccountInfo<'a>,
-    pub authority_signer_seeds: &'b [&'b [u8]],
-    pub token_program: AccountInfo<'a>,
-}
-
-pub struct TokenMintToParams<'a: 'b, 'b> {
-    pub mint: AccountInfo<'a>,
-    pub destination: AccountInfo<'a>,
-    pub amount: u64,
-    pub authority: AccountInfo<'a>,
-    pub authority_signer_seeds: &'b [&'b [u8]],
-    pub token_program: AccountInfo<'a>,
-}
-
-pub struct TokenBurnParams<'a: 'b, 'b> {
-    pub mint: AccountInfo<'a>,
-    pub source: AccountInfo<'a>,
-    pub amount: u64,
-    pub authority: AccountInfo<'a>,
-    pub authority_signer_seeds: &'b [&'b [u8]],
-    pub token_program: AccountInfo<'a>,
-}
